Added file path definition in config
Since files must be accessed as if in root, even in extensions; extensions can now define files in their directory and are loaded into a list. so now instead of accessing the files manually, extensions can do `filepaths['data.json]` for example and get the correct path to the file. this makes code more readable. Asker now uses this and documentation has been updated to mention it for developers.

a helper script file has also been started to give extensions and developers common functions